Add godot project with networking and FPS controls
This is the real initial commit with all the code I have until now.

Networking: A server and 2 clients connect and then control their
players. This is still buggy. For example, every client controls the
/same/ player.

FPS controls: WSAD, Jump, and Look with mouse are all implemented (Taken
from Caliou's demo). This includes collisions.