Count time on sides of objective; fix its physics
Changing the weight of the objective fixes the weird destroy-the-player
physics.

Also, count how long the objective has been tilted to the right or left,
and put counters there. Slow the game down and display who won when the
count exceeds 100% on either side. Like in overwatch, but with tilting.