[WIP] Network games, begin move players to game
The problem I just encountered is that I have to tell the players who
are currently skirmishing that they have a game. This means either:
 1. Telling the skirmish server to move them
 2. Keeping a connection with players
 3. Running the skirmish server on the matchmaker!
Three sounds nice because the matchmaker is sitting on this godot
instance, it might as well do something with it, right? But it's
essentially equivalent to 1, it just doesn't have to think about
networking because they're the same instance. So moving from 3 to 1 if I
realize it's a bad idea is easy, but 2 becomes pretty entrenched.

You're making some pretty big decisions here, take it slow.